Move files into common segmentation folder.
